trading-bot/requirements.txt
ccxt>=4.3.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pytz>=2024.1
